refactor: move shared notification logic into new file
