+++ lua/octo/pickers/telescope/previewers.lua
@@ -7,7 +7,7 @@ local utils = require "octo.utils"
 local previewers = require "telescope.previewers"
 local pv_utils = require "telescope.previewers.utils"
 local ts_utils = require "telescope.utils"
-local release = require "octo.release"
+local notifications = require "octo.notifications"
 local defaulter = ts_utils.make_default_callable
 local workflow_runs_previewer = require("octo.workflow_runs").previewer
 
@@ -114,68 +114,7 @@ local notification = defaulter(function(opts)
         local number = entry.value ---@type string
         local owner, name = utils.split_repo(entry.repo)
 
-        ---Fetches the data for the preview and then shows it with the given preview function
-        ---@param query string
-        ---@param fields table<string, string>
-        ---@param jq string
-        ---@param preview fun(obj: any, bufnr: integer): nil
-        local function fetch_and_preview(query, fields, jq, preview)
-          gh.api.graphql {
-            query = query,
-            fields = fields,
-            jq = jq,
-            opts = {
-              cb = gh.create_callback {
-                failure = vim.api.nvim_err_writeln,
-                success = function(output)
-                  if not vim.api.nvim_buf_is_loaded(bufnr) then
-                    return
-                  end
-
-                  local ok, obj = pcall(vim.json.decode, output)
-                  if not ok then
-                    utils.error("Failed to parse preview data: " .. vim.inspect(output))
-                    return
-                  end
-
-                  preview(obj, bufnr)
-                end,
-              },
-            },
-          }
-        end
-
-        ---@type string, table<string, string>, string, fun(obj: any, bufnr: integer): nil
-        local query, fields, jq, preview
-
-        if entry.kind == "issue" then
-          query = graphql("issue_query", owner, name, number, _G.octo_pv2_fragment)
-          fields = {}
-          jq = ".data.repository.issue"
-          preview = writers.issue_preview
-        elseif entry.kind == "pull_request" then
-          query = graphql("pull_request_query", owner, name, number, _G.octo_pv2_fragment)
-          fields = {}
-          jq = ".data.repository.pullRequest"
-          preview = writers.issue_preview
-        elseif entry.kind == "discussion" then
-          query = queries.discussion
-          fields = { owner = owner, name = name, number = number }
-          jq = ".data.repository.discussion"
-          preview = writers.discussion_preview
-        elseif entry.kind == "release" then
-          -- GraphQL only accepts tags and release notifications give back IDs
-          release.get_tag_from_release_id(entry, function(tag_name)
-            entry.tag_name = tag_name
-            query = queries.release
-            fields = { owner = owner, name = name, tag = tag_name }
-            jq = ".data.repository.release"
-            preview = writers.release_preview
-            fetch_and_preview(query, fields, jq, preview)
-          end)
-          return
-        end
-        fetch_and_preview(query, fields, jq, preview)
+        notifications.populate_preview_buf(bufnr, owner, name, number, entry.kind)
       end
     end,
   }
